WIP – move bundle-client to own file
