+++ lib/index.js
@@ -1,12 +1,10 @@
 module.exports = getHoodieServer
 
-var fs = require('fs')
-var path = require('path')
-
 var async = require('async')
 var hapi = require('hapi')
 var log = require('npmlog')
 
+var bundleClient = require('./bundle-client')
 var getConfig = require('./config')
 var registerPlugins = require('./plugins')
 var userDatabases = require('./utils/user-databases')
@@ -43,7 +41,7 @@ function getHoodieServer (options, callback) {
 
     async.parallel([
       registerPlugins.bind(null, server, config),
-      bundleHoodieClient.bind(null, config)
+      bundleClient.bind(null, config)
     ], function (error) {
       if (error) {
         return callback(error)
@@ -54,34 +52,6 @@ function getHoodieServer (options, callback) {
       server.plugins.account.api.accounts.on('remove', userDatabases.remove.bind(null, config, server))
 
       callback(null, server, config)
-    })
-  })
-}
-
-function bundleHoodieClient (config, callback) {
-  var hoodieClientModulePath = path.dirname(require.resolve('@hoodie/client/package.json'))
-  var hoodieClientPath = path.join(hoodieClientModulePath, 'dist/hoodie.js')
-  var bundleTargetPath = path.join(config.paths.data, 'client.js')
-
-  // https://github.com/hoodiehq/hoodie-client/issues/34
-  // var hoodieMinPath = path.join(hoodieClientModulePath, 'dist/hoodie.min.js')
-
-  log.silly('bundle', 'bundling ' + hoodieClientPath + ' into ' + bundleTargetPath)
-
-  var stream = fs.createReadStream(hoodieClientPath)
-  stream.pipe(fs.createWriteStream(bundleTargetPath))
-  stream.on('error', callback)
-  stream.on('end', function () {
-    // TODO: pass client configuration to constructor
-    fs.appendFile(bundleTargetPath, '\n\nhoodie = new Hoodie()', function (error) {
-      if (error) {
-        return callback(error)
-      }
-
-      log.silly('bundle', 'appended Hoodie init code to ' + bundleTargetPath)
-      log.info('bundle', 'bundled Hoodie client into ' + bundleTargetPath)
-
-      callback()
     })
   })
 }
