+++ lib/utils/user-databases.js
@@ -2,6 +2,10 @@
  * TODO: this functionality should be moved into hoodie-store-server. It should
  * expose an API at server.plugins.store.api the same way hoodie-account-server
  * does it: https://github.com/hoodiehq/hoodie-account-server/tree/master/api
+ *
+ * Right now it only works correctly if CouchDB is used for data. PouchDB
+ * creates Databases when used for the first time, but databases get currently
+ * not removed when a user account gets removed
  */
 module.exports = {
   add: addUserDatabase,
